+++ src/data/sample_logs.rs
@@ -96,7 +96,7 @@ impl SampleLog {
 
     /// Given a lower and upper limit, get the list of time starts and ends
     /// corresponding to the log filter.
-    pub fn to_time_ranges(&self, lower: f64, upper: f64) -> (Vec<usize>, Vec<usize>) {
+    pub fn to_time_ranges(&self, lower: f64, upper: f64) -> (Vec<u64>, Vec<u64>) {
         match self {
             SampleLog::I8(log) => log.to_time_ranges(&(lower as i8), &(upper as i8)),
             SampleLog::I16(log) => log.to_time_ranges(&(lower as i16), &(upper as i16)),
@@ -114,7 +114,7 @@ impl SampleLog {
     /// Given a list of filter starts and ends, return the sample log with filter applied.
     ///
     /// Assumes that start_times and end_times are sorted arrays.
-    pub fn apply_filters(&self, start_times: &[usize], end_times: &[usize]) -> SampleLog {
+    pub fn apply_filters(&self, start_times: &[u64], end_times: &[u64]) -> SampleLog {
         bind! {self, apply_filters(start_times, end_times)}
     }
 }
@@ -132,9 +132,9 @@ where
     T: PartialOrd,
 {
     /// Internal implementation of SampleLog.to_time_ranges
-    fn to_time_ranges(&self, lower: &T, upper: &T) -> (Vec<usize>, Vec<usize>) {
-        let mut starts = Vec::<usize>::new();
-        let mut ends = Vec::<usize>::new();
+    fn to_time_ranges(&self, lower: &T, upper: &T) -> (Vec<u64>, Vec<u64>) {
+        let mut starts = Vec::<u64>::new();
+        let mut ends = Vec::<u64>::new();
         let mut in_range: bool = false; // represents whether we are currently in the band
         for (index, value) in self.value.iter().enumerate() {
             if !in_range {
@@ -163,7 +163,7 @@ where
     T: Clone,
 {
     /// Internal implementation of SampleLog.apply_filters.
-    fn apply_filters(&self, start_times: &[usize], end_times: &[usize]) -> ValueLog<T> {
+    fn apply_filters(&self, start_times: &[u64], end_times: &[u64]) -> ValueLog<T> {
         // we use these indices to ignore overlaps in filters.
         //
         // Note that as the start_times and end_times are sorted, we will never get a scenario
@@ -388,7 +388,7 @@ mod tests {
         // 0  1  2  3  4  5  6  7  8  9  values
         //              ^------------... exclude
         let filter_starts = vec![0.45.to_ns()];
-        let filter_ends = vec![usize::MAX];
+        let filter_ends = vec![u64::MAX];
         let new_log = log.apply_filters(&filter_starts, &filter_ends);
 
         let expected_vals = Array1::<f64>::from_vec(vec![0., 1., 2., 3., 4.]);
